Relocate manubot/process/thumbnail.py
